fix: use createBrowserRouter
