+++ part2/part2_project/src/Components/sharedPage/sharedPageMain.jsx
